refactoring photo repository to service
